+++ test/functional/validation.acceptance.ts
@@ -29,8 +29,8 @@ suite("Acceptance validation tests (all sources)", () => {
                     includeRange: true
                 },
                 [
-                    "Warning: The parameter 'backupVaultRGIsNew' is never used. (ARM (Expressions)) [32,8-32,28]",
-                    "Warning: The parameter 'backupContainerName' is never used. (ARM (Expressions)) [47,8-47,29]"
+                    "Warning: The parameter 'backupVaultRGIsNew' is never used. (arm-template (expr)) [32,8-32,28]",
+                    "Warning: The parameter 'backupContainerName' is never used. (arm-template (expr)) [47,8-47,29]"
                 ])
     );