+++ test/functional/validation.expressions.test.ts
@@ -22,8 +22,8 @@ suite("Expression validation", () => {
                     includeSources: [sources.expressions]
                 },
                 [
-                    "Warning: The parameter 'backupVaultRGIsNew' is never used. (ARM (Expressions)) [32,8-32,28]",
-                    "Warning: The parameter 'backupContainerName' is never used. (ARM (Expressions)) [47,8-47,29]"
+                    "Warning: The parameter 'backupVaultRGIsNew' is never used. (arm-template (expr)) [32,8-32,28]",
+                    "Warning: The parameter 'backupContainerName' is never used. (arm-template (expr)) [47,8-47,29]"
                 ])
     );
 
@@ -36,10 +36,10 @@ suite("Expression validation", () => {
                 includeSources: [sources.expressions]
             },
             [
-                "Error: Undefined parameter reference: 'windowsOSVersion' (ARM (Expressions)) [69,26-69,44]",
-                "Error: Undefined variable reference: 'storageAccountType' (ARM (Expressions)) [116,35-116,55]",
-                "Warning: The parameter 'domainNamePrefix' is never used. (ARM (Expressions)) [4,4-4,22]",
-                "Warning: The variable 'osType' is never used. (ARM (Expressions)) [66,4-66,12]",
+                "Error: Undefined parameter reference: 'windowsOSVersion' (arm-template (expr)) [69,26-69,44]",
+                "Error: Undefined variable reference: 'storageAccountType' (arm-template (expr)) [116,35-116,55]",
+                "Warning: The parameter 'domainNamePrefix' is never used. (arm-template (expr)) [4,4-4,22]",
+                "Warning: The variable 'osType' is never used. (arm-template (expr)) [66,4-66,12]",
             ])
     );
 });